Enhance: Migrate pages folder to shared layout
