+++ client/src/_layout/pages/Home/index.tsx
@@ -1,0 +1,12 @@
+import GridMain from '@layout/components/GridMain';
+import Login from '../../../Login/components';
+
+function HomePage() {
+  return (
+    <GridMain>
+      <Login />
+    </GridMain>
+  );
+}
+
+export default HomePage;
