--- client/src/pages/Home/index.tsx
@@ -1,7 +1,0 @@
-import Login from '../../Login/components';
-
-function HomePage() {
-  return <Login />;
-}
-
-export default HomePage;
